requirements.txt
# Core Framework
fastapi
uvicorn

# Data & Utils
pydantic
python-dotenv

# OpenAI SDK (latest official)
openai>=1.42.0

# Vector Database (Pinecone)
pinecone-client>=4.1.0

# Optional - if you use pandas/numpy for data handling
numpy
pandas

# Logging and config
PyYAML

# Testing & Development
requests
httpx


